integração home com back + semifiltro

+++ src/main/java/br/com/farmaciasocialapi/resources/BaseRepository.java
@@ -2,6 +2,6 @@ package br.com.farmaciasocialapi.resources;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public abstract interface BaseRepository<E extends BaseEntity> extends JpaRepository<E, Integer>{
+public abstract interface BaseRepository<E extends BaseEntity> extends JpaRepository<E, Long>{
 
 }
